Break Parser common tests as per the source structure
