+++ test/lib/Streamly/Test/Data/Parser/CommonTests.hs
@@ -19,73 +19,43 @@
 -- Test.Parser test-suite.
 module Streamly.Test.Data.Parser.CommonTests (mainCommon, TestMode(..), takeWhileFailD) where
 
-import Control.Applicative ((<|>))
 import Control.Exception (displayException, try, evaluate, SomeException)
 import Data.List (isSuffixOf)
 import Streamly.Internal.Data.Fold (Fold(..))
-import Streamly.Internal.Data.MutByteArray (Unbox)
 import Streamly.Test.Common (listEquals, checkListEqual, chooseInt)
-import Streamly.Internal.Data.Parser (ParseErrorPos(..), Parser(..), Step(..), Initial(..), Final(..))
+import Streamly.Internal.Data.Parser (Parser(..), Step(..), Initial(..), Final(..))
 import Test.QuickCheck
        (arbitrary, forAll, elements, Property, property, listOf,
-        vectorOf, Gen, (.&&.), ioProperty)
-import Test.QuickCheck.Monadic (monadicIO, assert, run, PropertyM)
+        vectorOf, (.&&.), ioProperty)
+import Test.QuickCheck.Monadic (monadicIO, assert)
 
 import Prelude hiding (sequence, take, takeWhile)
 
-import qualified Control.Monad.Fail as Fail
 import qualified Data.List as List
 import qualified Prelude
 import qualified Streamly.Internal.Data.Stream as S
-import qualified Streamly.Internal.Data.Array as A
-import qualified Streamly.Internal.Data.Array.Generic as GA
 import qualified Streamly.Internal.Data.Fold as FL
 import qualified Streamly.Internal.Data.Parser as P
-import qualified Streamly.Internal.Data.ParserK as PK
-import qualified Streamly.Internal.Data.StreamK as K
 
 import Test.Hspec
 import Test.Hspec.QuickCheck
 
-#if MIN_VERSION_QuickCheck(2,14,0)
+import Control.Monad.Identity (Identity(runIdentity))
 
-import Test.QuickCheck (chooseAny)
-import Control.Monad.Identity (Identity(runIdentity, Identity))
-
-#else
-
-import System.Random (Random(random))
-import Test.QuickCheck.Gen (Gen(MkGen))
-
--- | Generates a random element over the natural range of `a`.
-chooseAny :: Random a => Gen a
-chooseAny = MkGen (\r _ -> let (x,_) = random r in x)
-
-#endif
-
-min_value :: Int
-min_value = 0
-
-mid_value :: Int
-mid_value = 5000
-
-max_value :: Int
-max_value = 10000
-
-max_length :: Int
-max_length = 1000
+import Streamly.Test.Data.Parser.CommonTestDriver
+    ( TestMode(..)
+    , ParserTestCase
+    , ParserTestCase_Temp
+    , runParserTC
+    , runParserTC_temp
+    , min_value
+    , mid_value
+    , max_value
+    , max_length
+    )
+import Streamly.Test.Data.Parser.CommonTypeTests (mainCommonType)
 
 -- Accumulator Tests
-
--- TODO: Replace ParserTestCase_Temp with ParserTestCase in all the test cases.
-
-type ParserTestCase a m b c =
-        (P.Parser a m b -> [a] -> m (Either ParseErrorPos b, [a])) -> c
-
-type ParserTestCase_Temp a m b c =
-        forall t. ([a] -> t)
-        -> (P.Parser a m b -> t -> m (Either ParseErrorPos b))
-        -> c
 
 fromFold :: ParserTestCase Int IO Int Property
 fromFold consumer =
@@ -97,43 +67,6 @@ fromFold consumer =
             case s1 of
                 Right o1 -> o1 == o2
                 Left _ -> False
-
-fromPure :: ParserTestCase Int Identity Int Property
-fromPure consumer =
-    forAll (chooseInt (min_value, max_value)) $ \x ->
-        case runIdentity $ consumer (P.fromPure x) [1 :: Int] of
-            (Right r, rest) -> r == x && rest == [1 :: Int]
-            (Left _, _) -> False
-
-fromEffect :: ParserTestCase Int Identity Int Property
-fromEffect consumer =
-    forAll (chooseInt (min_value, max_value)) $ \x ->
-        case runIdentity $ consumer (P.fromEffect $ return x) [1 :: Int] of
-            (Right r, rest) -> r == x && rest == [1 :: Int]
-            (Left _, _) -> False
-
-die :: ParserTestCase Int Identity Int Property
-die consumer =
-    property $
-    case runIdentity $ consumer (P.die "die test") [0 :: Int] of
-        (Right _, _) -> False
-        (Left _, rest) -> rest == [0 :: Int]
-
-dieM :: ParserTestCase Int Identity Int Property
-dieM consumer =
-    property $
-    case runIdentity $ consumer (P.dieM (Identity "die test")) [0 :: Int] of
-        (Right _, _) -> False
-        (Left _, rest) -> rest == [0 :: Int]
-
-parserFail :: ParserTestCase Int Identity Int Property
-parserFail consumer =
-    property $
-        case runIdentity $ consumer (Fail.fail err) [0 :: Int] of
-            (Right _, _) -> False
-            (Left (ParseErrorPos _ e), rest) -> err == e && rest == [0 :: Int]
-    where
-    err = "Testing MonadFail.fail."
 
 -- Element Parser Tests
 
@@ -570,35 +503,6 @@ wordBy producer consumer =
         let wrds = words lst
          in if wrds == [] && length lst > 0 then [""] else wrds
 
-splitWith :: ParserTestCase_Temp Int Identity (Int, Int) Property
-splitWith producer consumer =
-    forAll (listOf (chooseInt (0, 1))) $ \ls ->
-        case runIdentity $ consumer (P.splitWith (,) (P.satisfy (== 0)) (P.satisfy (== 1))) (producer ls) of
-            Right (result_first, result_second) -> case ls of
-                0 : 1 : _ -> (result_first == 0) && (result_second == 1)
-                _ -> False
-            Left _ -> case ls of
-                0 : 1 : _ -> False
-                _ -> True
-
-splitWithFailLeft :: ParserTestCase_Temp Int Identity (Int, Int) Property
-splitWithFailLeft producer consumer =
-    property (case runIdentity $ consumer (P.splitWith (,) (P.die "die") (P.fromPure (1 :: Int))) (producer [1 :: Int]) of
-        Right _ -> False
-        Left _ -> True)
-
-splitWithFailRight :: ParserTestCase_Temp Int Identity (Int, Int) Property
-splitWithFailRight producer consumer =
-    property (case runIdentity $ consumer (P.splitWith (,) (P.fromPure (1 :: Int)) (P.die "die")) (producer [1 :: Int]) of
-        Right _ -> False
-        Left _ -> True)
-
-splitWithFailBoth :: ParserTestCase_Temp Int Identity (Int, Int) Property
-splitWithFailBoth producer consumer =
-    property (case runIdentity $ consumer (P.splitWith (,) (P.die "die") (P.die "die")) (producer [1 :: Int]) of
-        Right _ -> False
-        Left _ -> True)
-
 -- teeWithPass :: ParserTestCase_Temp Int Identity Int Property
 -- teeWithPass producer consumer =
 --     forAll (chooseInt (min_value, max_value)) $ \n ->
@@ -749,74 +653,6 @@ someFail producer consumer =
     property (case runIdentity $ consumer (P.some (P.die "die") FL.toList) (producer [1 :: Int]) of
         Right _ -> False
         Left _ -> True)
-
--------------------------------------------------------------------------------
--- Instances
--------------------------------------------------------------------------------
-
-applicative :: ParserTestCase_Temp Int Identity ([Int], [Int]) Property
-applicative producer consumer =
-    forAll (listOf (chooseAny :: Gen Int)) $ \ list1 ->
-        forAll (listOf (chooseAny :: Gen Int)) $ \ list2 ->
-            let parser =
-                    (,)
-                        <$> P.fromFold (FL.take (length list1) FL.toList)
-                        <*> P.fromFold (FL.take (length list2) FL.toList)
-             in
-                case runIdentity $ consumer parser (producer $ list1 ++ list2) of
-                    Right (olist1, olist2) -> olist1 == list1 && olist2 == list2
-                    Left _ -> False
-
-sequence :: ParserTestCase_Temp Int IO [[Int]] Property
-sequence producer consumer =
-    forAll (vectorOf 11 (listOf (chooseAny :: Gen Int))) $ \ ins ->
-        let p xs = P.fromFold (FL.take (length xs) FL.toList)
-         in monadicIO $ do
-                outs <- run $
-                        consumer
-                            (Prelude.sequence $ fmap p ins)
-                            (producer $ concat ins)
-                return $
-                    case outs of
-                        Right ls -> ls == ins
-                        Left _ -> False
-
-altEOF1 :: ParserTestCase_Temp Int (PropertyM IO) Int Property
-altEOF1 producer consumer =
-    monadicIO $ do
-    s1 <- consumer
-        (P.satisfy (> 0) <|> return 66)
-        (producer ([]::[Int]))
-    return $
-        case s1 of
-            Right x -> x == 66
-            Left _ -> False
-
-altEOF2 :: ParserTestCase_Temp Int (PropertyM IO) [Int] Property
-altEOF2 producer consumer =
-    monadicIO $ do
-    s1 <- consumer
-        ((P.takeEQ 2 FL.toList) <|> (P.takeEQ 1 FL.toList))
-        (producer ([51]::[Int]))
-    return $
-        case s1 of
-            Right x -> x == [51]
-            Left _ -> False
-
-monad :: ParserTestCase_Temp Int (PropertyM IO) ([Int], [Int]) Property
-monad producer consumer =
-    forAll (listOf (chooseAny :: Gen Int)) $ \ list1 ->
-        forAll (listOf (chooseAny :: Gen Int)) $ \ list2 ->
-            let parser = do
-                    olist1 <- P.fromFold (FL.take (length list1) FL.toList)
-                    olist2 <- P.fromFold (FL.take (length list2) FL.toList)
-                    return (olist1, olist2)
-             in monadicIO $ do
-                    s <- consumer parser (producer $ list1 ++ list2)
-                    return $
-                        case s of
-                            Right (olist1, olist2) -> olist1 == list1 && olist2 == list2
-                            Left _ -> False
 
 takeEndBy1 :: ParserTestCase_Temp Int Identity [Int] Property
 takeEndBy1 producer consumer =
@@ -1114,84 +950,14 @@ takeWhileFailD predicate (Fold fstep finitial _ ffinal) =
 -- Main
 -------------------------------------------------------------------------------
 
-data TestMode
-    = TMParserStream
-    | TMParserKStreamK
-    | TMParserKStreamKChunks
-    | TMParserKStreamKChunksGeneric
-    deriving (Show)
-
-runParserTC :: (Unbox a, Monad m) => TestMode -> ParserTestCase a m b c -> c
-runParserTC tm runner =
-    case tm of
-        TMParserStream ->
-            runner $ \p -> mapMTup S.toList . S.parseBreakPos p . S.fromList
-        TMParserKStreamK ->
-            runner $ \p ->
-                mapMTup K.toList . K.parseBreakPos (PK.toParserK p) . K.fromList
-        TMParserKStreamKChunks ->
-            runner $ \p ->
-                mapMTup
-                    (fmap (concatMap A.toList) . K.toList)
-                        . A.parseBreakPos (A.toParserK p)
-                        . producerChunks A.fromList
-        TMParserKStreamKChunksGeneric ->
-            runner $ \p ->
-                mapMTup
-                    (fmap (concatMap GA.toList) . K.toList)
-                        . GA.parseBreakPos (GA.toParserK p)
-                        . producerChunks GA.fromList
-
-    where
-    mapMTup f tupM = do
-        (t, a) <- tupM
-        (t,) <$> f a
-
-    cSize = 50
-    -- Not using A.createOf here because of the MonadIO constraint
-    producerChunks fl =
-        K.fromStream
-             . S.groupsOf cSize (fl <$> FL.toList)
-             . S.fromList
-
-runParserTC_temp :: (Unbox a, Monad m) => TestMode -> ParserTestCase_Temp a m b c -> c
-runParserTC_temp tm runner =
-    case tm of
-        TMParserStream -> runner S.fromList S.parsePos
-        TMParserKStreamK -> runner K.fromList (K.parsePos . PK.toParserK)
-        TMParserKStreamKChunks ->
-            runner (producerChunks A.fromList) (A.parsePos . A.toParserK)
-        TMParserKStreamKChunksGeneric ->
-            runner
-                (producerChunks GA.fromList)
-                (GA.parsePos . GA.toParserK)
-
-    where
-    cSize = 50
-    -- Not using A.createOf here because of the MonadIO constraint
-    producerChunks fl =
-        K.fromStream
-             . S.groupsOf cSize (fl <$> FL.toList)
-             . S.fromList
-
 {-# NOINLINE mainCommon #-}
 mainCommon :: TestMode -> Spec
 mainCommon ptt = do
   describe (show ptt) $ do
-    describe "Instances" $ do
-        prop "applicative" $ runParserTC_temp ptt applicative
-        prop "Alternative: end of input 1" $ runParserTC_temp ptt altEOF1
-        prop "Alternative: end of input 2" $ runParserTC_temp ptt altEOF2
-        prop "monad" $ runParserTC_temp ptt monad
-        prop "sequence" $ runParserTC_temp ptt sequence
+    mainCommonType ptt
 
     describe "test for accumulator" $ do
         prop "P.fromFold FL.sum = FL.sum" $ runParserTC ptt fromFold
-        prop "fromPure value provided" $ runParserTC ptt fromPure
-        prop "fromPure monadic value provided" $ runParserTC ptt fromEffect
-        prop "fail err = Left (SomeException (ParseError err))" $ runParserTC ptt parserFail
-        prop "always fail" $ runParserTC ptt die
-        prop "always fail but monadic" $ runParserTC ptt dieM
 
     describe "test for element parser" $ do
         prop "peek = head with list length > 0" $ runParserTC ptt peekPass
@@ -1227,10 +993,6 @@ mainCommon ptt = do
         prop "groupByRolling" $ runParserTC_temp ptt groupByRolling
         prop "many (P.wordBy ' ') = words'" $ runParserTC_temp ptt wordBy
         -- prop "choice" choice
-        prop "parse 0, then 1, else fail" $ runParserTC_temp ptt splitWith
-        prop "fail due to die as left parser" $ runParserTC_temp ptt splitWithFailLeft
-        prop "fail due to die as right parser" $ runParserTC_temp ptt splitWithFailRight
-        prop "fail due to die as both parsers" $ runParserTC_temp ptt splitWithFailBoth
         -- prop "" teeWithPass
         -- prop "" teeWithFailLeft
         -- prop "" teeWithFailRight
